Add support for multiple docker containers
This commit allows us to create and use a separate docker container each time `tools/docker` gets changed.

Each image is tagged with the hash of the most recent commit that changed `tools/docker`. Let's call this `CONTIKER_TAG`

Travis tests will attempt to pull `contiker/contiki-ng:$CONTIKER_TAG`.

Then:
* If the pull succeeds then the PR / branch does not change `tools/docker`
* If `tools/docker` changed then the `docker pull` will fail. In this case, or if any other error occurs, we build a new docker image. This new docker image will be tagged `contiker/contiki-ng:$CONTIKER_TAG`

Lastly, if we are running on `contiki-ng/contiki-ng` and if the build was triggered by a branch, push `contiker/contiki-ng:$CONTIKER_TAG` to dockerhub.

To determine whether we are running on `contiki-ng/contiki-ng` (and not on a fork), we change the test to:

```
if [ $TRAVIS_REPO_SLUG == contiki-ng/contiki-ng ]
```

This is because `$TRAVIS_SECURE_ENV_VARS` can be true even outside `contiki-ng/contiki-ng`